[CHORE] #50 : 라벨 호출 Service 변경에 따라, Source 변경
