+++ client/app/src/main/java/com/example/healthc/data/source/product/ProductDataSourceImpl.kt
@@ -2,6 +2,7 @@ package com.example.healthc.data.source.product
 
 import com.example.healthc.data.model.remote.product.ProductItemResponse
 import com.example.healthc.data.model.remote.product.ProductResponse
+import com.example.healthc.data.model.remote.recipe.NutritionLabelResponse
 import com.example.healthc.data.service.ProductService
 import javax.inject.Inject
 
@@ -19,6 +20,16 @@ class ProductDataSourceImpl @Inject constructor(
     override suspend fun getProductList(query: String): Result<List<ProductItemResponse>>{
         return runCatching {
             service.getProductId(query = query).products
+        }.onFailure { e ->
+            e.printStackTrace()
+        }
+    }
+
+    override suspend fun getNutritionLabel(id: Int): Result<NutritionLabelResponse>{
+        return runCatching {
+            NutritionLabelResponse(
+                service.getNutritionLabel(id = id).bytes()
+            )
         }.onFailure { e ->
             e.printStackTrace()
         }
